extent wrapped by locality in span encoding: https://github.com/relaton/relaton-bib/issues/64
