+++ lib/metanorma/standoc/isodoc.rng
@@ -828,6 +828,9 @@
       </element>
     </define>
     <define name="BibDataExtensionType">
+      <optional>
+        <attribute name="schema-version"/>
+      </optional>
       <ref name="doctype"/>
       <optional>
         <ref name="docsubtype"/>
